자료형 변경

+++ SnakeApp/app/src/main/java/com/hello/myapp/snakeapp/framework/impl/AndroidGraphics.java
@@ -107,17 +107,17 @@ public class AndroidGraphics implements Graphics {
         srcRect.right = srcX + srcWidth - 1;
         srcRect.bottom = srcY + srcHeight - 1;
 
-        dstRect.left = x;
-        dstRect.top = y;
-        dstRect.right = x + srcWidth - 1;
-        dstRect.bottom = y + srcHeight - 1;
+        dstRect.left = (int) x;
+        dstRect.top = (int) y;
+        dstRect.right = (int) (x + srcWidth - 1);
+        dstRect.bottom = (int) (y + srcHeight - 1);
 
         canvas.drawBitmap(((AndroidPixmap) pixmap).bitmap, srcRect, dstRect,
                 null);
     }
     
     @Override
-    public void drawPixmap(Pixmap pixmap, int x, int y) {
+    public void drawPixmap(Pixmap pixmap, float x, float y) {
         canvas.drawBitmap(((AndroidPixmap)pixmap).bitmap, x, y, null);
     }
